requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
certifi==2016.9.26
configparser==3.5.0
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
gym==0.5.7
h5py==2.6.0
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
mock==2.0.0
nbconvert==5.0.0
nbformat==4.2.0
notebook==4.3.0
numpy==1.11.3
pandas==0.19.1
pandocfilters==1.4.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
protobuf==3.1.0
ptyprocess==0.5.1
pydot-ng==1.0.0
pyglet==1.2.4
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
tensorflow-gpu==0.12.0
terminado==0.6
testpath==0.3
Theano==0.8.2
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
